feature: added authentication functions to backend (#3)
In this commit I made sure to add authentication middleware that will
set and collect cookies when necessary. This will help us keep state of
when users are logged in or not. I also added unit tests for all new
functions and objects while also integrating new authentication method
to the registration route.