write unit tests for the login route (#4)
In this PR I focused on creating tests for the login route. This way @lcohen23 will be able to run npm run test in his terminal and be able to see if his code is up to par functionally. Pretty much what we are looking for in these tests are:

successful user login
login creates cookie
will not login user if it doesn't exist
will not login user if credentials don't line up
will not login if required fields are missing
will throw internal server error if for whatever reason there is a failed process
will not create cookie if login is unsuccessful